feature Paypal

+++ shop-online-angular/src/app/payment/payment.component.ts
@@ -7,7 +7,6 @@ import { OrderService } from '../service/order/order.service';
 import { CartService } from '../service/cart/cart.service';
 declare let paypal: any;
 declare let $: any;
-declare let checkPaypal: boolean;
 
 @Component({
   selector: 'app-payment',
@@ -18,6 +17,8 @@ export class PaymentComponent implements OnInit {
 
   paymentForm: any;
   orderItems: any;
+  total: any;
+  checkPaypal: boolean;
 
   constructor(
     private orderService: OrderService, 
@@ -32,10 +33,12 @@ export class PaymentComponent implements OnInit {
       phone: new FormControl(''),
       address: new FormControl('')
     });
+    this.checkPaypal = false;
   }
 
   ngOnInit() {
     this.orderItems = this.cartService.carts;
+    this.total = Math.floor(this.cartService.getTotal() / 21000);
     if (!this.orderItems[0]) {
       this.router.navigate(['/cart']);
     }
@@ -43,7 +46,8 @@ export class PaymentComponent implements OnInit {
       alert('You have to login!');
       this.router.navigate(['/cart']);
     }
-    $.getScript( 'https://www.paypalobjects.com/api/checkout.js', function() {
+    $.getScript( 'https://www.paypalobjects.com/api/checkout.js', $.proxy (() => {
+      var t = this.total;
       paypal.Button.render({
         env: 'sandbox', // sandbox | production
         client: {
@@ -51,44 +55,49 @@ export class PaymentComponent implements OnInit {
           production: '<insert production client id>'
         },
         commit: true,
-        payment: function(data, actions) {
+        payment: (data, actions) => {
           return actions.payment.create({
             payment: {
               transactions: [
               {
-                amount: { total: 1, currency: 'USD' }
+                amount: { total: t, currency: 'USD' }
               }
               ]
             }
           });
         },
-        onAuthorize: function(data, actions) {
-          return actions.payment.execute().then(function() {
-            console.log(data);
+        onAuthorize: (data, actions) => {
+          return actions.payment.execute().then(() => {
+            this.checkPaypal = true;
             window.alert('Payment Complete!');
           });
         }
       }, '#paypal-button-container');
-    });
+    }));
   }
 
   order(info) {
-    if (!localStorage.getItem('currentUser')) {
-      alert('You need to login');
-      return false;
+    if (this.checkPaypal) {
+      if (!localStorage.getItem('currentUser')) {
+        alert('You need to login');
+        return false;
+      }
+      let data = {
+        'name': info.name,
+        'phone': info.phone,
+        'address': info.address,
+        'orderItems': this.orderItems
+      }
+      this.orderService.createOrder(data).subscribe((a: any) => {
+        this.cartService.removeCart();
+        this.toasterService.pop('success','Order success!');
+        this.router.navigate(['history-orders']);
+      }, (err: any) => {
+        this.toasterService.pop('warning','Fail');
+      });
     }
-    let data = {
-      'name': info.name,
-      'phone': info.phone,
-      'address': info.address,
-      'orderItems': this.orderItems
+    else {
+      this.toasterService.pop('warning','You have to pay before order');
     }
-    this.orderService.createOrder(data).subscribe((a: any) => {
-      this.cartService.removeCart();
-      this.toasterService.pop('success','Order success!');
-      this.router.navigate(['history-orders']);
-    }, (err: any) => {
-      this.toasterService.pop('warning','Fail');
-    });
   }
 }
